Вторая часть: Небольшие дополнения и структура

+++ main.py
@@ -11,9 +11,14 @@ async def main():
     await bot_get_info()
     logger.bind(log_type="AEP", user="@Console").info(f"Начало запуска бота @{BotInfo.username}...")
 
+
+
+    # Нужно ли удалить веб-хук
+    if Permissions.delete_webhook:
+        await bot.delete_webhook()
+
     # Включение опроса бота
     await dp.start_polling(bot)
-    await bot.delete_webhook()
 
 
 # Вечная загрузка бота
